requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2
cryptography==2.9.2
Django==3.0.8
django-filter==2.3.0
django-phone-login==1.1.3
django-phonenumber-field==1.1.0
django-ratings==0.3.7
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-sendsms==0.2.3
django-twilio==0.13.0.post1
djangorestframework==3.11.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
openapi-codec==1.3.2
pandas==1.0.5
phonenumbers==8.12.7
phonenumberslite==8.12.7
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
twilio==6.44.2
uritemplate==3.0.1
urllib3==1.25.9
